io: move io_rpc_trans_cb, rpc_trans_writer, io_handle_write, io_do_tls to handlers.c
All four functions are now backend-agnostic since rpc_trans_writer
routes submissions through the io_resubmit_write primitive each
backend provides (see commit 1).  Move them to the shared
lib/io/handlers.c alongside the existing accept/connect handlers.

write.c shrinks to ~200 lines containing only two liburing
functions: io_request_write_op (creates a fresh ic, submits write)
and io_resubmit_write (submits the next chunk of an existing ic).
Everything else -- orchestration, TLS glue, completion handling --
is now in handlers.c and shared between backends.

Remove the io_rpc_trans_cb and io_handle_write stubs from
backend_kqueue.c in the same commit so there is no duplicate
symbol.  The kqueue main-loop dispatcher (write_and_dispatch) now
calls the real io_handle_write from handlers.c, which will reach
io_resubmit_write (currently -ENOSYS stub on kqueue -- commit 6
implements it via kevent).

No semantic change on Linux: the moved code is byte-identical to
the pre-move definitions, with adjustments only for the shared
file's include list (adds openssl/bio.h, openssl/err.h,
openssl/ssl.h, reffs/tls.h, time.h).

FreeBSD PR #7 precondition: the TLS branch in rpc_trans_writer
(ci->ci_ssl && ci->ci_tls_enabled) is unreachable because
io_tls_init_server_context is not called from the kqueue
io_handler_init.  PR #8 will port TLS and must address the
EVFILT_WRITE collision that arises when io_do_tls submits outside
the write gate -- see PR #7 addendum B2.
